steps back and forward
reverting a lot of the code for the main parent component. i wanted it to initialize some of its content only the one time, and i couldnt find a way to write that without a constructor tbqh. im being obsessive about performance, and i reverted the context provider setup from before. now its a lot cleaner since some of the functions i was using before were one liners that i can now skip over the bloated .bind(this) syntax of.

on the topic of being obsessive, the parent component is doing an extra render of the child output component when the userInput is changed. from reading the docs, ive seen them mention a parent rendering and a child rendering as acceptable, im trying to make them independent, but i dont think my instincts are correct with how ive set this up. i just need to minimize re-render work done by the parser in the output. like memoization and check if the input has changed sort of thing. i think. or maybe the one snippet of docs that's led me to this conclusion isn't appropriate for my case....but idk im spending a ton of time on something thats 1 or 2ms tops.
one thing i want to do is change the logic of the parser function so it loops through and parses characters individually. the dictionary had a single line that was checking if the string was valid, and that check is out of line with what the dictionary should have for its logical structure. looping through the characters and passing them each to the dictionary from the parser seems like a better solution to write than the alternative i had, which was to do some filtering and mapping before and after interacting with the dictionary.
i feel like there's more to change with how im passing props into the two child function components, stylistically. im going to try and find some established typescript-react repositories for my own reference.
i still want to look into other advanced react things, namely useMemo, as it sounds fitting. then i need to move on to more functionality and stop fretting over some of these details. everything's working, i've just been prodding around a lot in the muck. which is good for learning, but ive been in the same spot for a bit now, that's all