some cleanup!
found a better name for the dictionary method. coming up with a name was a good idea, because it made me realize it should be called "getOrMemoizeDict". wait a minute...that should be two methods then! refactored! cleanliness ahoy!
oh and just edited/deleted some todo comments that no longer made sense or couldve been better worded for future me